Differentiate between source and canvas
Previously, the x and y values of `VectorRGBXY`s created when
loading an image didn't line up with the actual pixel values.
This is because the image was resized to `IMAGE_WIDTH` before the
ImageData was extracted.  This was preventing updating just the
colors of the pointCloud from an external script which knew the
pixels of the image but not `IMAGE_WIDTH`.

This was fixed by keeping two different instances of ImageData in
scope: one for the source, and one for after the image was resized.
The mask in `showPointsInImage()` is set to be the size of the
source so that parts of it can be revealed using the source's
pixel locations, and then the mask is resized using the 9-param
overload of `context.drawImage()`.